examples/deno/localconvert/pyproject.toml
[tool.poetry]
name = "localconvert"
version = "0.1.0"
description = ""
authors = ["DKL <dkl@zurich.ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow-jxl-plugin = "^1.3.2"
docling = "^2.28.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
